Add tsb subscribe command and allow for naming remotes
The subscription feature allows for users to subscribe to branches.
Ths subscribe command will add the branch to the patches config. Running
tsb update after a subscription is made, adds any commits to the patches file
between the origin head and the branch.

The repos yaml file now allows extras to be either strings specifying the path
or objects containing a path and a name. This allows users to give meaningful
names to the remotes that are added via the extra feature when running tsb fetch.

Documentation is updated with this commmit as well